refactor(GitHub): enhance integration test base class
- Expanded the `GithubIntegrationTest` class to include methods for creating mock responses for organizations, repositories, and pull requests, improving the test setup for GitHub integration tests.
- Removed redundant code from `test_integration_resync.py` by utilizing the shared base class, streamlining the test structure.
- Updated documentation within the base class to clarify its purpose and usage for future test implementations.
